+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.5.50 (Pending):
 * Update EntitySubscriptionImpl to remove type map when empty.
-* Ensure that entities are unloaded  when subscriptions are removed as part of disconnect.
+* Ensure that entities are unloaded from EntityRepository and listeners in the change broker
+  are removed when subscriptions are removed as part of disconnect() method in AbstractDataLoaderService.
 * Update EntityRepositoryDebugger to add methods to support debugging of subscriptions.
 
 ## 0.5.49: